+++ rust/main/hyperlane-base/src/settings/parser/mod.rs
@@ -42,11 +42,21 @@ const DEFAULT_CHUNK_SIZE: u32 = 1999;
 #[serde(transparent)]
 pub struct RawAgentConf(Value);
 
+fn agent_name_to_default_rpc_consensus_type(agent_name: &str) -> String {
+    match agent_name {
+        "validator" => "quorum".to_string(),
+        "relayer" => "fallback".to_string(),
+        "scraper" => "fallback".to_string(),
+        _ => "fallback".to_string(),
+    }
+}
+
 impl FromRawConf<RawAgentConf, Option<&HashSet<&str>>> for Settings {
     fn from_config_filtered(
         raw: RawAgentConf,
         cwp: &ConfigPath,
         filter: Option<&HashSet<&str>>,
+        agent_name: &str,
     ) -> Result<Self, ConfigParsingError> {
         let mut err = ConfigParsingError::default();
 
@@ -91,16 +101,12 @@ impl FromRawConf<RawAgentConf, Option<&HashSet<&str>>> for Settings {
             .and_then(parse_signer)
             .end();
 
-        let default_rpc_consensus_type = p
-            .chain(&mut err)
-            .get_opt_key("defaultRpcConsensusType")
-            .parse_string()
-            .unwrap_or("fallback");
+        let default_rpc_consensus_type = agent_name_to_default_rpc_consensus_type(agent_name);
 
         let chains: HashMap<String, ChainConf> = raw_chains
             .into_iter()
             .filter_map(|(name, chain)| {
-                parse_chain(chain, &name, default_rpc_consensus_type)
+                parse_chain(chain, &name, default_rpc_consensus_type.as_str())
                     .take_config_err(&mut err)
                     .map(|v| (name, v))
             })
@@ -426,6 +432,7 @@ impl FromRawConf<RawAgentSignerConf> for SignerConf {
         raw: RawAgentSignerConf,
         cwp: &ConfigPath,
         _filter: (),
+        _agent_name: &str,
     ) -> ConfigResult<Self> {
         parse_signer(ValueParser::new(cwp.clone(), &raw.0))
     }