+++ typescript/sdk/src/metadata/agentConfig.ts
@@ -296,12 +296,6 @@ export const AgentConfigSchema = z.object({
   defaultSigner: AgentSignerSchema.optional().describe(
     'Default signer to use for any chains that have not defined their own.',
   ),
-  defaultRpcConsensusType: z
-    .nativeEnum(RpcConsensusType)
-    .describe(
-      'The default consensus type to use for any chains that have not defined their own.',
-    )
-    .optional(),
   log: z
     .object({
       format: z
@@ -585,6 +579,5 @@ export function buildAgentConfig(
 
   return {
     chains: chainConfigs,
-    defaultRpcConsensusType: RpcConsensusType.Fallback,
   };
 }